# docs/contracts/maekon-web.v1.openapi.yaml
openapi: 3.0.3
info:
  title: Maekon Local Web API
  version: "v1"
  description: |
    Auto-generated from docs/contracts/http-interface-manifest.v1.json.
    Request/response bodies are real schemars-generated DTO schemas
    (full disclosure; the API is loopback-only + token-gated). Cross-crate
    maekon-core field types are rendered as opaque objects (build containment).
  x-maekon-document-version: 10
  x-maekon-updated-at: "2026-07-23"
  x-maekon-routes-file: "crates/maekon-web/src/routes.rs"
  x-maekon-contracts-crate: "crates/maekon-api-contracts"
servers:
  - url: /
paths:
...
  "/api/sync/setup":
    get:
      tags:
        - sync_setup
      operationId: sync_setup_get_api_sync_setup
      summary: "GET /api/sync/setup"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSetupStatus'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - sync_setup
      operationId: sync_setup_post_api_sync_setup
      summary: "POST /api/sync/setup"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncSetupRequest'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSetupStatus'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse: {"properties":{"code":{"type":"string"},"error":{"type":"string"},"status":{"format":"uint16","maximum":65535,"minimum":0,"type":"integer"}},"required":["code","error","status"],"type":"object"}
...